requirements.txt
streamlit
langchain
langchain-community
langgraph
google-genai
qdrant-client
python-dotenv
faiss-cpu
PyPDF2
ollama